requirements.txt
autopep8==1.5.6
click==7.1.2
Faker==6.6.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycodestyle==2.7.0
python-dateutil==2.8.1
python-dotenv==0.17.0
six==1.15.0
SQLAlchemy==1.3.23
text-unidecode==1.3
toml==0.10.2
Werkzeug==1.0.1
WTForms==2.3.3
